+++ intelmq/bots/parsers/cymru/parser_cap_program.py
@@ -28,9 +28,12 @@ MAPPING_STATIC = {'bot': {
 }
 MAPPING_COMMENT = {'bruteforce': ('classification.identifier', 'protocol.application'),
                    'phishing': ('source.url', )}
-PROTOCOL_MAPPING = {'6': 'tcp',  # TODO: use getent in harmonization
+PROTOCOL_MAPPING = {  # TODO: use getent in harmonization
+                    '1': 'icmp',
+                    '6': 'tcp',
+                    '11': 'nvp-ii',
                     '17': 'udp',
-                    '1': 'icmp'}
+                    }
 BOGUS_HOSTNAME_PORT = re.compile('hostname: ([^:]+)port: ([0-9]+)')
 DESTINATION_PORT_NUMBERS_TOTAL = re.compile(r' \(total_count:\d+\)$')