+++ intelmq/tests/bots/parsers/cymru/test_cap_program_new.py
@@ -180,10 +180,16 @@ EVENTS = [{'time.source': '2019-03-22T11:18:52+00:00',
            'classification.type': 'phishing',
            'classification.identifier': 'phishing',
            },
+          {'classification.type': 'scanner',
+           'classification.identifier': 'darknet',
+           'protocol.transport': 'nvp-ii',
+           'destination.port': 0,
+           'time.source': '2020-01-10T09:17:17+00:00',
+           },
           ]
 # The number of events a single line in the raw data produces
 NUM_EVENTS = [1, 1, 1, 1, 1, 2, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,
-              1, 10, 1, 1]
+              1, 10, 1, 1, 1]
 RAWS = []
 for i, line in enumerate(RAW_LINES[3:]):
     for count in range(NUM_EVENTS[i]):