+++ CHANGELOG.md
@@ -75,15 +75,19 @@ CHANGELOG
 ### Configuration
 
 ### Core
+- `intelmq.lib.upgrades`:
+  - Harmonization upgrade: Also check and update regular expressions
 
 ### Development
 
 ### Harmonization
+- `protocol.transport`: Adapt regular expression to allow the value `nvp-ii` (protocol 11).
 
 ### Bots
 #### Collectors
 
 #### Parsers
+- `intelmq.bots.parser.cymru.parser_cap_program`: Support for protocol 11 (`nvp-ii`).
 
 #### Experts